fix(precompiles): derive multisig eligibility from registry
